requirements.txt
django
djangorestframework
djangorestframework_simplejwt
python-dotenv
drf-spectacular
drf-spectacular[sidecar]